requirements.txt
Django==5.2.3
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv
site-packages (1.1.1)